backend/intern_portal_backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.0
gunicorn>=20.1
